requirements.txt
polars>=1.6.0
duckdb>=1.1.0
pyarrow>=16.0.0
streamlit>=1.37.0
pydeck>=0.9.1
h3>=4.1.0
huggingface-hub>=0.24.0
python-dotenv>=1.0.1
openai>=1.40.0
pytest>=8.0.0
matplotlib>=3.8.0
tabulate>=0.9.0
reportlab>=4.0.0
openpyxl>=3.1.0
shapely>=2.0.0
pyshp>=2.3.0
